fix: correct clippy and format issues.

+++ tracer/src/instruction/format/format_u.rs
@@ -45,7 +45,7 @@ impl InstructionFormat for FormatU {
 			} | // imm[63:32] = [31]
 			((word as u64) & 0xfffff000)
                 // imm[31:12] = [31:12]
-            ) as u64,
+            ),
         }
     }
